Final backend refactor
