+++ src/main/java/ecdar/abstractions/Query.java
@@ -1,13 +1,19 @@
 package ecdar.abstractions;
 
+import EcdarProtoBuf.QueryProtos;
+import com.google.gson.JsonParser;
 import ecdar.Ecdar;
 import ecdar.backend.*;
 import ecdar.controllers.EcdarController;
+import ecdar.utility.UndoRedoStack;
+import ecdar.utility.helpers.StringValidator;
 import ecdar.utility.serialize.Serializable;
 import com.google.gson.JsonObject;
 import javafx.application.Platform;
 import javafx.beans.property.*;
+import javafx.collections.ObservableList;
 
+import java.util.NoSuchElementException;
 import java.util.function.Consumer;
 
 public class Query implements Serializable {
@@ -56,11 +62,15 @@ public class Query implements Serializable {
         }
     };
 
-    public Query(final String query, final String comment, final QueryState queryState) {
+    public Query(final String query, final String comment, final QueryState queryState, final Engine engine) {
         this.query.set(query);
         this.comment.set(comment);
         this.queryState.set(queryState);
-        setEngine(BackendHelper.getDefaultEngine());
+        setEngine(engine);
+    }
+
+    public Query(final String query, final String comment, final QueryState queryState) {
+        this(query, comment, queryState, BackendHelper.getDefaultEngine());
     }
 
     public Query(final JsonObject jsonElement) {
@@ -143,6 +153,109 @@ public class Query implements Serializable {
 
     public Consumer<Exception> getFailureConsumer() {
         return failureConsumer;
+    }
+
+    /**
+     * Executes the query
+     */
+    public void execute() throws NoSuchElementException {
+        if (getQueryState().equals(QueryState.RUNNING) || !StringValidator.validateQuery(getQuery()))
+            return;
+
+        if (getQuery().isEmpty()) {
+            setQueryState(QueryState.SYNTAX_ERROR);
+            addError("Query is empty");
+            return;
+        }
+
+        setQueryState(QueryState.RUNNING);
+        errors().set("");
+
+        getEngine().enqueueRequest(this, this::handleQueryResponse, this::handleQueryBackendError);
+    }
+
+    private void handleQueryResponse(QueryProtos.QueryResponse value) {
+        // If the query has been cancelled, ignore the result
+        if (getQueryState() == QueryState.UNKNOWN) return;
+
+        if (value.hasRefinement() && value.getRefinement().getSuccess()) {
+            setQueryState(QueryState.SUCCESSFUL);
+            getSuccessConsumer().accept(true);
+        } else if (value.hasConsistency() && value.getConsistency().getSuccess()) {
+            setQueryState(QueryState.SUCCESSFUL);
+            getSuccessConsumer().accept(true);
+        } else if (value.hasDeterminism() && value.getDeterminism().getSuccess()) {
+            setQueryState(QueryState.SUCCESSFUL);
+            getSuccessConsumer().accept(true);
+        } else if (value.hasComponent()) {
+            setQueryState(QueryState.SUCCESSFUL);
+            getSuccessConsumer().accept(true);
+            JsonObject returnedComponent = (JsonObject) JsonParser.parseString(value.getComponent().getComponent().getJson());
+            addGeneratedComponent(new Component(returnedComponent));
+        } else {
+            setQueryState(QueryState.ERROR);
+            getSuccessConsumer().accept(false);
+        }
+    }
+
+    private void handleQueryBackendError(Throwable t) {
+        // If the query has been cancelled, ignore the error
+        if (getQueryState() == QueryState.UNKNOWN) return;
+
+        // Each error starts with a capitalized description of the error equal to the gRPC error type encountered
+        String errorType = t.getMessage().split(":\\s+", 2)[0];
+
+        if ("DEADLINE_EXCEEDED".equals(errorType)) {
+            setQueryState(QueryState.ERROR);
+            getFailureConsumer().accept(new BackendException.QueryErrorException("The engine did not answer the request in time"));
+        } else {
+            try {
+                setQueryState(QueryState.ERROR);
+                getFailureConsumer().accept(new BackendException.QueryErrorException("The execution of this query failed with message:" + System.lineSeparator() + t.getLocalizedMessage()));
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    private void addGeneratedComponent(Component newComponent) {
+        Platform.runLater(() -> {
+            newComponent.setTemporary(true);
+
+            ObservableList<Component> listOfGeneratedComponents = Ecdar.getProject().getTempComponents(); // ToDo NIELS: Refactor
+            Component matchedComponent = null;
+
+            for (Component currentGeneratedComponent : listOfGeneratedComponents) {
+                int comparisonOfNames = currentGeneratedComponent.getName().compareTo(newComponent.getName());
+
+                if (comparisonOfNames == 0) {
+                    matchedComponent = currentGeneratedComponent;
+                    break;
+                } else if (comparisonOfNames < 0) {
+                    break;
+                }
+            }
+
+            if (matchedComponent == null) {
+                UndoRedoStack.pushAndPerform(() -> { // Perform
+                    Ecdar.getProject().getTempComponents().add(newComponent);
+                }, () -> { // Undo
+                    Ecdar.getProject().getTempComponents().remove(newComponent);
+                }, "Created new component: " + newComponent.getName(), "add-circle");
+            } else {
+                // Remove current component with name and add the newly generated one
+                Component finalMatchedComponent = matchedComponent;
+                UndoRedoStack.pushAndPerform(() -> { // Perform
+                    Ecdar.getProject().getTempComponents().remove(finalMatchedComponent);
+                    Ecdar.getProject().getTempComponents().add(newComponent);
+                }, () -> { // Undo
+                    Ecdar.getProject().getTempComponents().remove(newComponent);
+                    Ecdar.getProject().getTempComponents().add(finalMatchedComponent);
+                }, "Created new component: " + newComponent.getName(), "add-circle");
+            }
+
+            Ecdar.getProject().addComponent(newComponent);
+        });
     }
 
     @Override
