+++ src/main/java/ecdar/controllers/EngineOptionsDialogController.java
@@ -5,7 +5,7 @@ import com.google.gson.JsonParser;
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXRippler;
 import ecdar.Ecdar;
-import ecdar.abstractions.Engine;
+import ecdar.backend.Engine;
 import ecdar.backend.BackendException;
 import ecdar.backend.BackendHelper;
 import ecdar.presentations.EnginePresentation;
@@ -73,7 +73,7 @@ public class EngineOptionsDialogController implements Initializable {
 
             // Close all engine connections to avoid dangling engine connections when port range is changed
             try {
-                Ecdar.getBackendDriver().clear();
+                BackendHelper.clearEngineConnections(); // ToDO NIELS: Only clear affected engines/connections
             } catch (BackendException e) {
                 // ToDO NIELS: Handle exceptions from clearing backend
             }
