+++ src/test/java/ecdar/backend/QueryTest.java
@@ -1,0 +1,18 @@
+package ecdar.backend;
+
+import ecdar.abstractions.Query;
+import ecdar.abstractions.QueryState;
+import org.junit.jupiter.api.Test;
+
+import static org.mockito.Mockito.*;
+
+public class QueryTest {
+    @Test
+    public void testExecuteQuery() {
+        Engine e = mock(Engine.class);
+        Query q = new Query("refinement: (Administration || Machine || Researcher) <= Spec", "", QueryState.UNKNOWN, e);
+        q.execute();
+
+        verify(e, times(1)).enqueueRequest(eq(q), any(), any());
+    }
+}
